More configuration for corporate mode

+++ sh/setup_debian.sh
@@ -5,6 +5,10 @@ if [[ -z "${dev_user// }" ]]; then
     echo "A value must be supplied for the dev_user."
     exit 1
 fi
+corporate_mode=$2
+if [[ -z "${corporate_mode// }" ]]; then
+    echo "A value must be supplied for corporate mode (true or false)"
+fi
 
 function install_ansible_prerequisites() {
     apt-get update -y
@@ -12,17 +16,23 @@ function install_ansible_prerequisites() {
 }
 
 function install_ansible() {
-    pip install --upgrade pyasn1 setuptools --cert /usr/local/share/ca-certificates/corp.crt
-    pip install ansible==2.4.0.0 --cert /usr/local/share/ca-certificates/corp.crt
+    if [[ "$corporate_mode" == "true" ]]; then
+        pip install --upgrade pyasn1 setuptools --cert /usr/local/share/ca-certificates/corp.crt
+        pip install ansible==2.4.0.0 --cert /usr/local/share/ca-certificates/corp.crt
+    else
+        pip install --upgrade pyasn1 setuptools
+        pip install ansible==2.4.0.0
+    fi
 }
 
 function setup_git() {
     # Git comes installed with Ubuntu (at least the Vagrant box, anyway), and for the proxy
     # environment, it needs to be removed to be replaced with a different version that uses
     # the openssl library rather than gnutls.
-    # This needs an extra conditional to detect proxy mode.
-    if [[ ! -f "/home/$dev_user/.git_openssl_mod_installed" ]]; then
-        apt-get remove git -y
+    if [[ "$corporate_mode" == "true" ]]; then
+        if [[ ! -f "/home/$dev_user/.git_openssl_mod_installed" ]]; then
+            apt-get remove git -y
+        fi
     fi
 }
